Change wallet's stock tx id to Integer for consistency

+++ src/main/java/seng468/scalability/models/entity/WalletTX.java
@@ -1,7 +1,6 @@
 package seng468.scalability.models.entity;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
@@ -11,14 +10,14 @@ import jakarta.persistence.Table;
 @Table(name = "`Wallet Transactions`")
 public class WalletTX {
     @Id
-    private long walletTXId;
+    private Integer walletTXId;
     private long stockTXId;
     private boolean isDebit;
     private int amount;
     private LocalDateTime timestamp;
 
 
-    public WalletTX(long walletTXId, long stockTXId, boolean isDebit, int amount) {
+    public WalletTX(Integer walletTXId, long stockTXId, boolean isDebit, int amount) {
         this.walletTXId = walletTXId;
         this.stockTXId = stockTXId;
         this.isDebit = isDebit;
@@ -30,7 +29,7 @@ public class WalletTX {
 
     }
 
-    public long getWalletTXId() {
+    public Integer getWalletTXId() {
         return this.walletTXId;
     }
